third_party/WebKit/LayoutTests/platform/win/external/wpt/mediacapture-streams/MediaDevices-enumerateDevices.https-expected.txt

This is a testharness.js-based test.
PASS mediaDevices.enumerateDevices() is present and working on navigator 
Harness: the test ran to completion.
